Utils file split into a preliminary package
This is a draft. As there might be an additional file created to handle
"data_structure" utilities.
